UICCAI-864: cx-test domain breakout part 1
