--- cx-test/src/main/kotlin/io/nuvalence/cx/tools/cxtest/assertion/ContextAwareAssertionError.kt
@@ -1,4 +1,0 @@
-package io.nuvalence.cx.tools.cxtest.assertion
-
-class ContextAwareAssertionError(message: String?, val sourceId: String, val sourceLocator: Any?) : Exception(message) {
-}
