--- cx-test/src/main/kotlin/io/nuvalence/cx/tools/cxtest/testsource/FormatReader.kt
@@ -1,7 +1,0 @@
-package io.nuvalence.cx.tools.cxtest.testsource
-
-import io.nuvalence.cx.tools.cxtest.model.test.TestScenario
-
-interface FormatReader {
-    fun read(range: String) : List<TestScenario>
-}
