--- cx-test/src/main/kotlin/io/nuvalence/cx/tools/cxtest/util/Mappings.kt
@@ -1,13 +1,0 @@
-package io.nuvalence.cx.tools.cxtest.util
-
-val ACTION_MAPPINGS = mapOf(
-    "[Call begins]#en" to "Hello",
-    "[Call begins]#es" to "Hola",
-    "[No Input]#en" to "",
-    "[No Input]#es" to ""
-)
-
-val LANGUAGE_MAPPINGS = mapOf(
-    "en" to "en-US",
-    "es" to "es-ES"
-)
