+++ cx-test/src/main/resources/default.properties
@@ -1,4 +1,4 @@
 credentialsUrl=file:///Users/ginoelteng/git/creds/nuv_client_secret_ccai-sheets.json
-agentPath=projects/dol-uisim-ccai-dev-app/locations/global/agents/7d8acdfb-e622-4b98-bcd0-3ba2a92fde4e
-spreadsheetId=1wa6OsN5FUORHHPLGdRn8fYJ8TV9bA16pyzTr2Q5meVs
+agentPath=projects/dol-uisim-ccai-dev-app/locations/global/agents/ee332e8e-de2c-4d65-884f-405682eeaf3c
 dfcxEndpoint=dialogflow.googleapis.com:443
+dfcxTagFilter=#UAT
