+++ dol-ui-ccai-domain/src/main/kotlin/gov/ny/dol/ui/ccai/dfcx/domain/assertion/ContextAwareAssertionError.kt
@@ -1,0 +1,4 @@
+package gov.ny.dol.ui.ccai.dfcx.domain.assertion
+
+class ContextAwareAssertionError(message: String?, val sourceId: String, val sourceLocator: Any?) : Exception(message) {
+}
