+++ dol-ui-ccai-domain/src/main/kotlin/gov/ny/dol/ui/ccai/dfcx/domain/testsource/FormatReader.kt
@@ -1,0 +1,7 @@
+package gov.ny.dol.ui.ccai.dfcx.domain.testsource
+
+import gov.ny.dol.ui.ccai.dfcx.domain.model.test.TestScenario
+
+interface FormatReader {
+    fun read(range: String) : List<TestScenario>
+}
