+++ dol-ui-ccai-domain/src/main/kotlin/gov/ny/dol/ui/ccai/dfcx/domain/util/Mappings.kt
@@ -1,0 +1,13 @@
+package gov.ny.dol.ui.ccai.dfcx.domain.util
+
+val ACTION_MAPPINGS = mapOf(
+    "[Call begins]#en" to "Hello",
+    "[Call begins]#es" to "Hola",
+    "[No Input]#en" to "",
+    "[No Input]#es" to ""
+)
+
+val LANGUAGE_MAPPINGS = mapOf(
+    "en" to "en-US",
+    "es" to "es-ES"
+)
