+++ settings.gradle.kts
@@ -8,3 +8,4 @@ include("cx-variations")
 include("cx-suggestions")
 include("cx-intentgen")
 include("cx-test-sync")
+include("dol-ui-ccai-domain")
